Renamed io.anserini.end2end -> io.anserini.integration
