--- src/test/java/io/anserini/end2end/README.md
@@ -1,2 +1,0 @@
-This test is to automatically test indexing, searching and evaluation from end to end,
-making sure that they work as expected. This is more like an automatic regression test.
